+++ packages/tracing/tests/index.test.ts
@@ -10,7 +10,7 @@ describe("InMemoryTracer:", () => {
   });
 
   describe("rootSpan function:", () => {
-    test("it should create basic `InMemorySpan`", async (done) => {
+    test("it should create basic `InMemorySpan`", async () => {
       expect.assertions(2);
 
       spanNames.forEach((spanName) => {
@@ -23,26 +23,22 @@ describe("InMemoryTracer:", () => {
       expect(spans.length).toEqual(1);
 
       expect(spans[0].name).toBe("first-span");
-
-      done();
     });
   });
 
   describe("span function:", () => {
-    test("it should create `InMemorySpan`", async (done) => {
+    test("it should create `InMemorySpan`", async () => {
       expect.assertions(8);
 
       spanNames.forEach((spanName, index) =>
         tracer.span(spanName, async (result) => {
-          expect(result.id).toBe(index + 1);
+          expect(result.id).toBe((index + 1).toString());
           expect(result.name).toBe(spanName);
         }),
       );
-
-      done();
     });
 
-    test("it should store spans", async (done) => {
+    test("it should store spans", async () => {
       expect.assertions(1);
 
       for await (const spanName of spanNames) {
@@ -50,13 +46,11 @@ describe("InMemoryTracer:", () => {
       }
 
       expect(tracer.spans.length).toEqual(4);
-
-      done();
     });
   });
 
   describe("searchSpanByName function:", () => {
-    test("it should return all the span named as the argument", async (done) => {
+    test("it should return all the span named as the argument", async () => {
       expect.assertions(3);
 
       for await (const spanName of spanNames) {
@@ -70,8 +64,6 @@ describe("InMemoryTracer:", () => {
       spans.forEach((span) => {
         expect(span.name).toBe("second-span");
       });
-
-      done();
     });
   });
 });
@@ -84,7 +76,7 @@ describe("InMemorySpan:", () => {
   });
 
   describe("setAttribute function:", () => {
-    test("it should add a single attribute to the span", async (done) => {
+    test("it should add a single attribute to the span", async () => {
       expect.assertions(1);
 
       await tracer.span("test-span", async (span) => {
@@ -97,13 +89,11 @@ describe("InMemorySpan:", () => {
       const expectedAttributes = { "test-attribute": "[REDACTED]" };
 
       expect(spanAttributes).toStrictEqual(expectedAttributes);
-
-      done();
     });
   });
 
   describe("setDisclosedAttribute function:", () => {
-    test("it should add a single attribute to the span", async (done) => {
+    test("it should add a single attribute to the span", async () => {
       expect.assertions(1);
 
       await tracer.span("test-span", async (span) => {
@@ -116,86 +106,28 @@ describe("InMemorySpan:", () => {
       const expectedAttributes = { "test-disclosed-attribute": "testValue" };
 
       expect(spanAttributes).toStrictEqual(expectedAttributes);
-
-      done();
-    });
-  });
-
-  describe("setAttributes function:", () => {
-    test("it should add a hash of attributes to the span", async (done) => {
-      expect.assertions(1);
-
-      const attributeHash = {
-        "test-attribute-1": "testValue1",
-        "test-attribute-2": "testValue2",
-        "test-attribute-3": "testValue3",
-      };
-
-      await tracer.span("test-span", async (span) => {
-        span.setAttributes(attributeHash);
-
-        return span;
-      });
-
-      const spanAttributes = tracer.searchSpanByName("test-span")[0].attributes;
-
-      expect(spanAttributes).toStrictEqual(attributeHash);
-
-      done();
     });
   });
 
   describe("addEvent function:", () => {
-    test("it should return the current in memory span", async (done) => {
-      expect.assertions(1);
-
-      tracer.span("test-span", async (span) => {
-        expect(span.addEvent()).toBe(span);
-      });
-
-      done();
-    });
-  });
-
-  describe("setStatus function:", () => {
-    test("it should return the current in memory span", async (done) => {
-      expect.assertions(1);
-
-      tracer.span("test-span", async (span) => {
-        expect(span.addEvent()).toBe(span);
-      });
-
-      done();
-    });
-  });
-
-  describe("isRecording function:", () => {
-    test("it should return true", async (done) => {
-      expect.assertions(1);
+    test("it should add events to the span", async () => {
+      expect.assertions(5);
+      const expectedTime = new Date();
 
       await tracer.span("test-span", async (span) => {
-        expect(span.isRecording()).toBe(true);
+        expect(span.addEvent("event")).toBe(span);
+        span.addEvent("event2", expectedTime);
       });
 
-      done();
-    });
-  });
+      const spanEvents = tracer.searchSpanByName("test-span")[0].events;
 
-  describe("updateName function:", () => {
-    test("it should the name of the span", async (done) => {
-      expect.assertions(2);
-
-      await tracer.span("test-span", async (span) => span);
-
-      const span = tracer.searchSpanByName("test-span")[0];
-
-      expect(span.name).toStrictEqual("test-span");
-
-      span.updateName("new-name");
-
-      expect(span.name).toStrictEqual("new-name");
-
-      done();
+      expect(spanEvents.length).toBe(2);
+      expect(spanEvents[0]).toMatchObject({ name: "event" });
+      expect(spanEvents[0].attributes.startTime).toBeDefined();
+      expect(spanEvents[1]).toMatchObject({
+        name: "event2",
+        attributes: { startTime: expectedTime.toString() },
+      });
     });
   });
 });